arrumado possível overflow de pontos

+++ src/include/Tipos.h
@@ -119,7 +119,7 @@ typedef struct Jogador {
     int quantidadePulos;
     int quantidadeMaxPulos;
 
-    int quantidadePontos;
+    long int quantidadePontos;
     float quantidadeTempo;
     int quantidadeAneis;
     int quantidadeVidas;
